electron/python-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
diffusers==0.25.0
transformers==4.37.0
torch==2.2.0
pillow==10.2.0
accelerate==0.26.0
